Long

+++ outdoor_segmentation/pcseg/model/segmentor/voxel/epcl_hd/epcl_outdoor_seg_exp5.py
@@ -691,7 +691,7 @@ class EPCLOutdoorSegHD(BaseSegmentor):
             coords, not_outlier, tuple_feat = self.hd_model.clean_z(batch_dict['lidar'].C, tuple_feat)
             hv, sim, pred_label = self.hd_model.forward(tuple_feat, coords = coords)
             print(pred_label.shape)
-            pred_label_z = torch.zeros((batch_dict['lidar'].C.shape[0]))
+            pred_label_z = torch.zeros((batch_dict['lidar'].C.shape[0]), dtype=torch.int64)
             pred_label_z[not_outlier] = pred_label
             print(pred_label_z.shape)
             for idx in range(invs.C[:, -1].max() + 1):
